mostly fixing the watchdog autostart problem...
turns out, that when a function is wrapped in PicoWeb `@app` decorator you
can't call it normally. Create a wrapper for `set_auto` so that `main()` can
call it to put the watchdog into automatic mode at startup, and so that the
web api can also use it.

the clock_correcting_timer callback function was wrong

improve the logic for counting the number of resets, both automatic and
manually triggered.